Fixed typescript typings

+++ src/controllers/me.ts
@@ -4,6 +4,8 @@ import joi from 'joi';
 import { User } from '../database/entities/user';
 import { userInputValidationSchema } from '../database/validators';
 import { hashPassword } from './auth';
+import { Article } from '../database/entities/article';
+import { Audiofile } from '../database/entities/audiofile';
 
 const MESSAGE_ME_NOT_FOUND = 'Your account is not found. This could happen when your account is (already) deleted.';
 const MESSAGE_ME_DELETED = 'Your account is deleted. This cannot be undone.';
@@ -21,18 +23,30 @@ export const findCurrentUser = async (req: Request, res: Response) => {
 
 export const findAllArticles = async (req: Request, res: Response) => {
   const userId = req.user.id;
-  const userRepository = getRepository(User);
+  const articleRepository = getRepository(Article);
 
-  const { articles } = await userRepository.findOne(userId, { relations: ['articles'] });
+  const articles = await articleRepository.find({
+    where: {
+      user: {
+        id: userId
+      }
+    }
+  });
 
   return res.json(articles);
 };
 
 export const findAllAudiofiles = async (req: Request, res: Response) => {
   const userId = req.user.id;
-  const userRepository = getRepository(User);
+  const audiofileRepository = getRepository(Audiofile);
 
-  const { audiofiles } = await userRepository.findOne(userId, { relations: ['audiofiles'] });
+  const audiofiles = await audiofileRepository.findOne({
+    where: {
+      user: {
+        id: userId
+      }
+    }
+  });
 
   return res.json(audiofiles);
 };
@@ -89,6 +103,8 @@ export const deleteCurrentUser = async (req: Request, res: Response) => {
   }
 
   const user = await userRepository.findOne(userId);
+
+  if (!user) return res.status(400).json({ message: 'User not found!' });
 
   await userRepository.remove(user);
